pyproject.toml
[project]
name = "backend-coreappointment"
version = "0.1.0"
description = "SaaS Modular de Agendamiento - Backend Core"
authors = [
    {name = "hernan.harco", email = "hernan.harco@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0.0"
dependencies = [
    "fastapi (>=0.129.0,<0.130.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "langgraph (>=1.0.9,<2.0.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "pytz (>=2025.2,<2026.0)",
    "thefuzz (>=0.22.1,<0.23.0)",
    "openai (>=2.21.0,<3.0.0)",
    "rich (>=14.3.3,<15.0.0)",
    "langgraph-cli[inmem] (>=0.4.12,<0.5.0)",
    "langchain-openai (>=1.1.10,<2.0.0)",
    "langgraph-api (>=0.7.50,<0.8.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "dateparser (>=1.3.0,<2.0.0)",
    "langgraph-checkpoint-postgres (>=3.0.4,<4.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "psycopg-pool (>=3.3.0,<4.0.0)",
    "langgraph-sdk (>=0.3.9,<0.4.0)",
    "pgvector (>=0.4.2,<0.5.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "ipykernel (>=7.2.0,<8.0.0)",
    "poethepoet (>=0.41.0,<0.42.0)"
]

[tool.poetry]
package-mode = false

# --- 🚀 COMANDOS PERSONALIZADOS (Tipo pnpm scripts) ---
[tool.poe.tasks]
dev = "uvicorn app.main:app --host 0.0.0.0 --port 8002 --reload --reload-dir app --reload-include '*.py' --reload-include '*.env'" # Ejemplo: poetry run poe dev
test = "pytest tests/ -v"
setup-dev = "./setup.sh development"
setup-prod = "./setup.sh production"
test-agent = "python tests/test_agent.py"
init-checkpoints = "python -m app.agents.core.init_checkpoints"

# 📦 COMANDOS DE BASE DE DATOS (ALEMBIC)
db-plan = "alembic revision --autogenerate -m"   # Ejemplo: poetry run poe db-plan "mensaje"
db-migrate = "alembic upgrade head"
db-rollback = "alembic downgrade -1"
db-history = "alembic history --verbose"

# para activar el LangSmith: poetry run langgraph dev